+++ packages/components/__generated__/MembersListFragment.graphql.ts
@@ -1,0 +1,215 @@
+/**
+ * @generated SignedSource<<4538f21a2c43f1393fe96f42f844c6a0>>
+ * @lightSyntaxTransform
+ * @nogrep
+ */
+
+/* tslint:disable */
+
+/* eslint-disable */
+// @ts-nocheck
+import { ReaderFragment, RefetchableFragment } from 'relay-runtime'
+import { FragmentRefs } from 'relay-runtime'
+
+export type MembersListFragment$data = {
+  readonly id: string
+  readonly participants:
+    | {
+        readonly edges: ReadonlyArray<
+          | {
+              readonly node:
+                | {
+                    readonly profile:
+                      | {
+                          readonly id: string
+                          readonly ' $fragmentSpreads': FragmentRefs<'ProfileItemFragment'>
+                        }
+                      | null
+                      | undefined
+                    readonly role: string | null | undefined
+                  }
+                | null
+                | undefined
+            }
+          | null
+          | undefined
+        >
+        readonly pageInfo: {
+          readonly endCursor: string | null | undefined
+          readonly hasNextPage: boolean
+        }
+        readonly totalCount: number | null | undefined
+      }
+    | null
+    | undefined
+  readonly ' $fragmentType': 'MembersListFragment'
+}
+export type MembersListFragment$key = {
+  readonly ' $data'?: MembersListFragment$data
+  readonly ' $fragmentSpreads': FragmentRefs<'MembersListFragment'>
+}
+
+const node: ReaderFragment = (function () {
+  var v0 = ['participants'],
+    v1 = {
+      alias: null,
+      args: null,
+      kind: 'ScalarField',
+      name: 'id',
+      storageKey: null,
+    }
+  return {
+    argumentDefinitions: [
+      {
+        defaultValue: 5,
+        kind: 'LocalArgument',
+        name: 'count',
+      },
+      {
+        defaultValue: null,
+        kind: 'LocalArgument',
+        name: 'cursor',
+      },
+    ],
+    kind: 'Fragment',
+    metadata: {
+      connection: [
+        {
+          count: 'count',
+          cursor: 'cursor',
+          direction: 'forward',
+          path: v0 /*: any*/,
+        },
+      ],
+      refetch: {
+        connection: {
+          forward: {
+            count: 'count',
+            cursor: 'cursor',
+          },
+          backward: null,
+          path: v0 /*: any*/,
+        },
+        fragmentPathInResult: ['node'],
+        operation: require('./ChatRoomParticipantsPaginationQuery.graphql'),
+        identifierInfo: {
+          identifierField: 'id',
+          identifierQueryVariableName: 'id',
+        },
+      },
+    },
+    name: 'MembersListFragment',
+    selections: [
+      v1 /*: any*/,
+      {
+        alias: 'participants',
+        args: null,
+        concreteType: 'ChatRoomParticipantConnection',
+        kind: 'LinkedField',
+        name: '__ChatRoom_participants_connection',
+        plural: false,
+        selections: [
+          {
+            alias: null,
+            args: null,
+            concreteType: 'ChatRoomParticipantEdge',
+            kind: 'LinkedField',
+            name: 'edges',
+            plural: true,
+            selections: [
+              {
+                alias: null,
+                args: null,
+                concreteType: 'ChatRoomParticipant',
+                kind: 'LinkedField',
+                name: 'node',
+                plural: false,
+                selections: [
+                  {
+                    alias: null,
+                    args: null,
+                    concreteType: 'Profile',
+                    kind: 'LinkedField',
+                    name: 'profile',
+                    plural: false,
+                    selections: [
+                      v1 /*: any*/,
+                      {
+                        args: null,
+                        kind: 'FragmentSpread',
+                        name: 'ProfileItemFragment',
+                      },
+                    ],
+                    storageKey: null,
+                  },
+                  {
+                    alias: null,
+                    args: null,
+                    kind: 'ScalarField',
+                    name: 'role',
+                    storageKey: null,
+                  },
+                  {
+                    alias: null,
+                    args: null,
+                    kind: 'ScalarField',
+                    name: '__typename',
+                    storageKey: null,
+                  },
+                ],
+                storageKey: null,
+              },
+              {
+                alias: null,
+                args: null,
+                kind: 'ScalarField',
+                name: 'cursor',
+                storageKey: null,
+              },
+            ],
+            storageKey: null,
+          },
+          {
+            alias: null,
+            args: null,
+            concreteType: 'PageInfo',
+            kind: 'LinkedField',
+            name: 'pageInfo',
+            plural: false,
+            selections: [
+              {
+                alias: null,
+                args: null,
+                kind: 'ScalarField',
+                name: 'hasNextPage',
+                storageKey: null,
+              },
+              {
+                alias: null,
+                args: null,
+                kind: 'ScalarField',
+                name: 'endCursor',
+                storageKey: null,
+              },
+            ],
+            storageKey: null,
+          },
+          {
+            alias: null,
+            args: null,
+            kind: 'ScalarField',
+            name: 'totalCount',
+            storageKey: null,
+          },
+        ],
+        storageKey: null,
+      },
+    ],
+    type: 'ChatRoom',
+    abstractKey: null,
+  }
+})()
+
+;(node as any).hash = '4aad0edd43aec1e87daea42f647a59d3'
+
+export default node