+++ packages/components/modules/messages/CreateChatRoomList/ChatRoomListItem/types.ts
@@ -1,8 +1,6 @@
-import { Dispatch, SetStateAction } from 'react'
-
 import { ProfileItemFragment$key } from '../../../../__generated__/ProfileItemFragment.graphql'
 
 export interface ChatRoomListItemProps {
   profile: ProfileItemFragment$key
-  setIsInExistingChatRoomsView: Dispatch<SetStateAction<boolean>>
+  onChatCreation: () => void
 }