+++ packages/components/modules/messages/GroupDetails/GroupDetailsHeader/types.ts
@@ -1,0 +1,5 @@
+export interface GroupDetailsHeaderProps {
+  onBackButtonClicked: () => void
+  onEditButtonClicked?: () => void
+  shouldDisplayEditButton: boolean
+}