+++ packages/components/modules/messages/MessagesList/index.tsx
@@ -63,7 +63,7 @@ const MessagesList: FC<MessagesListProps> = ({
   const firstUnreadMessageId = useMemo(getFirstUnreadMessageId, [selectedRoom, currentProfile])
 
   useEffect(() => {
-    if (room?.unreadMessages?.count !== 0) {
+    if (room?.unreadMessages?.count !== 0 || room?.unreadMessages.markedUnread) {
       commitMutation({
         variables: {
           input: {
@@ -120,7 +120,6 @@ const MessagesList: FC<MessagesListProps> = ({
       hasNext,
       room?.isGroup,
       firstUnreadMessageId,
-      room?.participants?.totalCount,
       MessagesGroup,
       MessagesGroupProps,
     ],