+++ packages/components/modules/messages/graphql/queries/MembersList.ts
@@ -1,0 +1,25 @@
+import { graphql } from 'react-relay'
+
+export const MembersListFragment = graphql`
+  fragment MembersListFragment on ChatRoom
+  @argumentDefinitions(cursor: { type: "String" }, count: { type: "Int", defaultValue: 5 })
+  @refetchable(queryName: "ChatRoomParticipantsPaginationQuery") {
+    id
+    participants(first: $count, after: $cursor) @connection(key: "ChatRoom_participants") {
+      edges {
+        node {
+          profile {
+            id
+            ...ProfileItemFragment
+          }
+          role
+        }
+      }
+      pageInfo {
+        hasNextPage
+        endCursor
+      }
+      totalCount
+    }
+  }
+`