+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseapp-frontend/graphql
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/utils@3.1.1
+
 ## 1.2.0
 
 ### Minor Changes